Move content type methods
